Revert "#6 Run circuits parellely"
This reverts commit 060f523b4abbe8e4ff3b518f04bfadc01bc0c054.
