+++ src/quanvolutional_filter.py
@@ -74,6 +74,7 @@ class QuanvolutionalFilter:
 
         # Transpile the circuit.
         self.simulator = qiskit_aer.AerSimulator()
+        self.transpiled_circuit = qiskit.transpile(self.circuit, self.simulator)
 
     def __build_initial_circuit(self):
         """Build the initial ciruit."""
@@ -158,7 +159,7 @@ class QuanvolutionalFilter:
         for gate, qubits in self.selected_gates:
             self.circuit.append(gate, qubits)
 
-    def load_data(self, encoded_data: np.ndarray):
+    def __load_data(self, encoded_data: np.ndarray):
         """Load the data to the cirucit.
 
         :param np.ndarray encoded_data: _description_
@@ -178,16 +179,51 @@ class QuanvolutionalFilter:
         except:
             print(self.circuit.draw())
 
-    def run(self, encoded_data: np.ndarray, shots: int) -> dict:
-        ####
-        # This method is for a single use not for QuanvolutionalLayer.
-        ####
-        # Load the given encoded data to the circuit.
-        self.load_data(encoded_data=encoded_data)
+    def run(self, data: np.ndarray, shots: int) -> dict:
+        # Encode the data.
+        encoded_data = QuanvolutionalFilter.encode_with_threshold(data)
+
+        # Load the data to the circuit.
+        self.__load_data(encoded_data)
 
         # Run the circuit.
-        transpiled_circuit = qiskit.transpile(self.circuit, self.simulator)
-        result = self.simulator.run(transpiled_circuit, shots=shots).result()
-        counts = result.get_counts(transpiled_circuit)
+        result = self.simulator.run(self.transpiled_circuit, shots=shots).result()
+        counts = result.get_counts(self.transpiled_circuit)
 
-        return counts
+        # Decode the data.
+        decoded_data = QuanvolutionalFilter.decode_by_summing_ones(counts)
+
+        return decoded_data
+
+    @staticmethod
+    def encode_with_threshold(data: np.ndarray, threshold: float = 1) -> np.ndarray:
+        """Encode the given data according to the threshold. This method is suggested in the original paper.
+
+        :param np.ndarray data: original data
+        :param float threshold: threshold to encode
+        :return np.ndarray: encoded data
+        """
+        flatten_data = data.flatten()
+        encode_flags = np.where(flatten_data >= threshold, 1, 0).astype(np.float64)
+        quantum_state = 1
+        for encode_flag in encode_flags:
+            encoded_state = np.array([1, 0]) if encode_flag == 0 else np.array([0, 1])
+            quantum_state = np.kron(quantum_state, encoded_state)
+
+        return quantum_state
+
+    @staticmethod
+    def decode_by_summing_ones(counts: dict) -> int:
+        """Decode the measured result to the number of ones in the result.
+
+        :param dict counts: result of running the circuit
+        :return int: the number of ones in the most likely result
+        """
+        # Sort the resuly by the frequency.
+        sorted_counts = dict(sorted(counts.items(), key=lambda item: -item[1]))
+        # Get the most likely result.
+        most_likely_result = list(sorted_counts.keys())[0]
+        # Count the number of ones.
+        num_ones = most_likely_result.count("1")
+
+        return num_ones
